Refactorization
- moving functions after merge of master with new module strucuter
- binary arithmetic operations add, sub, div, mul, pow from opertations.py to arithmetics.py
- relational operations eq, equal, ne, le, lt, ge, gt, min, max from operations.py to new file relations.py
- adjusting unit tests
